Try loading products from DB.
