+++ includes/Modules/Subscribe_With_Google.php
@@ -207,6 +207,11 @@ final class Subscribe_With_Google extends Module
 				array(
 					'src'           => $base_url . 'js/googlesitekit-subscribe-with-google-bulk-edit.js',
 					'dependencies'  => array(
+						'googlesitekit-modules-subscribe-with-google',
+						'googlesitekit-api',
+						'googlesitekit-data',
+						'googlesitekit-datastore-site',
+						'googlesitekit-modules',
 						'googlesitekit-vendor',
 					),
 					'load_contexts' => array( Asset::CONTEXT_ADMIN_POSTS ),
